+++ src/commands/mod.rs
@@ -1,0 +1,4 @@
+pub mod extract;
+pub mod merge;
+pub mod shell_completion;
+pub mod sort;